+++ src/components/Menu/ToolbarMenu/ToolbarMenuComponent.tsx
@@ -161,6 +161,8 @@ export class ToolbarMenuComponent extends React.Component {
                                     icon={widgetConfig.isCustomIcon ? <CustomIcon icon={widgetConfig.icon as CustomIconName} /> : (widgetConfig.icon as IconName)}
                                     id={`${trimmedStr}Button`} // id particularly is for drag source in WidgetStore
                                     onClick={widgetConfig.onClick}
+                                    draggable={true}
+                                    onDragStart={(e: React.DragEvent) => WidgetsStore.Instance.handleToolbarWidgetDragStart(e, widgetConfig.widgetConfig)}
                                 />
                             </Tooltip>
                         );
@@ -230,7 +232,7 @@ export class ToolbarMenuComponent extends React.Component {
                             }
                             position={Position.BOTTOM}
                         >
-                            <AnchorButton icon={"console"} onClick={() => appStore.dialogStore.showDialog(DialogId.Snippet)} active={dialogStore.dialogVisible.get(DialogId.Snippet)} />
+                            <AnchorButton icon={"console"} onClick={() => appStore.dialogStore.showDialog(DialogId.Snippet)} active={dialogStore.dialogVisible.get(DialogId.Snippet)} data-testid={DialogId.Snippet + "-button"} />
                         </Tooltip>
                     )}
                 </ButtonGroup>